Fix format-string warning for 32-bit compilation

+++ src/binary-reader.cc
@@ -629,7 +629,7 @@ Result BinaryReader::ReadAddress(Address* out_value,
                                  const char* desc) {
   ERROR_UNLESS(memory < memories.size(),
                "load/store memory %u out of range %lu", memory,
-               memories.size());
+               static_cast<unsigned long>(memories.size()));
   if (memories[memory].is_64) {
     return ReadU64Leb128(out_value, desc);
   } else {
